Sliders now only modify ratios by +-25%

+++ nebree8.go
@@ -18,9 +18,9 @@ const (
 )
 
 type Ingredient struct {
-	Parts int    `json:"parts,omitempty"`
-	Drops int    `json:"drops,omitempty"`
-	Name  string `json:"name"`
+	Parts float32 `json:"parts,omitempty"`
+	Drops int     `json:"drops,omitempty"`
+	Name  string  `json:"name"`
 }
 
 type Order struct {
